Expose additional trigger information from build cause
With this, it should be possible to use multiple different triggers and to
execute different steps based on the particular trigger. It also provides a way
to notify the user who triggered the build via GitHub comments.

The full comment is exposed so that trigger commands could be parameterized.
E.g. extracting the module name from `test module-x`, which matches a generic
`test.*` trigger.